refactor: split vault/record.rs and vault/mod.rs tests into sub-modules
record.rs (591 lines) → vault/record/ directory:
- kind.rs      — RecordKind
- label.rs     — RecordLabel + is_forbidden_control
- payload.rs   — RecordPayload + RecordPayloadEncrypted
- aggregate.rs — Record struct + impl (new, rehydrate, with_updated_*)
- tests.rs     — #[cfg(test)] tests
- mod.rs       — module declarations + pub use re-exports

vault/mod.rs tests (296 lines) → vault/tests.rs:
- vault/mod.rs trimmed to 227 lines (well under 500)
- all test helpers and TC-U07 tests moved to vault/tests.rs

All public API paths unchanged; shikomi_core re-exports unaffected.
