+++ crates/shikomi-core/src/vault/record/kind.rs
@@ -1,0 +1,14 @@
+//! レコード種別。
+
+// -------------------------------------------------------------------
+// RecordKind
+// -------------------------------------------------------------------
+
+/// レコードの種別。
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum RecordKind {
+    /// テキストレコード（URL・メモ等、機密度が低い）。
+    Text,
+    /// シークレットレコード（パスワード・鍵等、機密度が高い）。
+    Secret,
+}
